managed_context_ops: carve anchors.rs (pure move)
The context-rewind anchor catalog: rollout user turns, managed-context
edit records, anchor list/inspect/scan over the session rollout,
compact-catalog serialization, outcome keys, rollout usage extraction,
anchor matching and summaries, primer-fact pruning, and pressure
bands; 32 tests and their eight rollout-fixture helpers ride along.
